+++ stowPackages/neovim/.config/nvim/init.lua
@@ -1,0 +1,14 @@
+-- ##########################
+-- Main Config File
+-- ##########################
+
+require "core.options"			-- core options
+require "core.keymaps"			-- custom keymaps
+require "core.colorscheme"		-- custom colorscheme
+
+-- LSP Configurations
+require "lsp"					-- LSP Configuration and Setup
+
+-- Plugin Configurations
+require "plugins"				-- Plugins Configuration and Setup
+