+++ stowPackages/neovim/.config/nvim/lua/plugins/config/suda.lua
@@ -1,0 +1,2 @@
+
+vim.g.suda_smart_edit = 1			-- auto sudos a file when opening a file with sudo perms needed