+++ stowPackages/neovim/.config/nvim/plugin_configs/chadtree.lua
@@ -1,6 +1,0 @@
-
-
-chadtree_settings = { ["view.open_direction"] = "right" }
-vim.api.nvim_set_var("chadtree_settings", chadtree_settings)
-
-