+++ CHANGELOG.md
@@ -10,9 +10,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### TODO:
 
 - Add Recommended VS Code/ium Extensions (in separate directory)
-- Add Vim-Plugged auto-install script
+~~ add Vim-Plugged auto-install script ~~ (DONE)
 - Add Zsh auto-install script
 - Make AwesomeWM theme.lua
+
+## [2.2.0] - 2022-07-08
+
+- Minor Neovim Full Lua Config Upgrade
+
+### Updated
+- neovim
+	- Updated Neovim config to full lua!
+	- Added additional plugins
 
 ## [2.1.12] - 2022-06-14